% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respect.R
\name{respect}
\alias{respect}
\title{A function to show respect}
\usage{
respect(x)
}
\arguments{
\item{x}{input string}
}
\value{
input string preceeded by "I respect"
}
\description{
A function to show respect
}
\examples{
respect::respect('trees')
}
